+++ lib/ts/types.ts
@@ -17,24 +17,32 @@ export type InputType = {
     apiDomain: string;
     apiBasePath?: string;
     sessionScope?: string;
-    refreshAPICustomHeaders?: any;
-    signoutAPICustomHeaders?: any;
     sessionExpiredStatusCode?: number;
     autoAddCredentials?: boolean;
     isInIframe?: boolean;
     cookieDomain?: string;
+    preAPIHook?: (context: {
+        action: "SIGN_OUT" | "REFRESH_SESSION";
+        requestInit: RequestInit;
+        url: string;
+    }) => Promise<{ url: string; requestInit: RequestInit }>;
+    onHandleEvent?: (context: { action: "SIGN_OUT" | "REFRESH_SESSION" | "UNAUTHORISED" }) => void;
 };
 
 export type NormalisedInputType = {
     apiDomain: string;
     apiBasePath: string;
     sessionScope: string;
-    refreshAPICustomHeaders?: any;
-    signoutAPICustomHeaders?: any;
     sessionExpiredStatusCode: number;
     autoAddCredentials: boolean;
     isInIframe: boolean;
     cookieDomain: string | undefined;
+    preAPIHook: (context: {
+        action: "SIGN_OUT" | "REFRESH_SESSION";
+        requestInit: RequestInit;
+        url: string;
+    }) => Promise<{ url: string; requestInit: RequestInit }>;
+    onHandleEvent: (context: { action: "SIGN_OUT" | "REFRESH_SESSION" | "UNAUTHORISED" }) => void;
 };
 
 export type PreAPIHookFunction = (context: {
@@ -43,9 +51,9 @@ export type PreAPIHookFunction = (context: {
 }) => Promise<{ url: string; requestInit: RequestInit }>;
 
 export interface RecipeInterface {
-    addFetchInterceptors: (env: any, originalFetch: any, config: NormalisedInputType) => Promise<void>;
+    addFetchInterceptorsAndReturnModifiedFetch: (originalFetch: any, config: NormalisedInputType) => typeof fetch;
 
-    addAxiosInterceptors: (axiosInstance: any, config: NormalisedInputType) => Promise<void>;
+    addAxiosInterceptors: (axiosInstance: any, config: NormalisedInputType) => void;
 
     getUserId: (config: NormalisedInputType) => Promise<string>;
 
@@ -54,8 +62,4 @@ export interface RecipeInterface {
     doesSessionExist: (config: NormalisedInputType) => Promise<boolean>;
 
     signOut: (config: NormalisedInputType) => Promise<void>;
-
-    // saveSessionFromResponse: (context: { requestInit: RequestInit; url: string; response: Response }) => Promise<void>;
-
-    // attachSessionToRequest: PreAPIHookFunction;
 }